+++ packages/core/src/renderables/Diff.regression.test.ts
@@ -7,7 +7,7 @@ import { ManualClock } from "../testing/manual-clock.js"
 import { MockTreeSitterClient } from "../testing/mock-tree-sitter-client.js"
 import type { SimpleHighlight } from "../lib/tree-sitter/types.js"
 import { BoxRenderable } from "./Box.js"
-import type { CodeRenderable } from "./Code.js"
+import { settleDiffHighlighting } from "./__tests__/renderable-test-utils.js"
 
 let currentRenderer: TestRenderer
 let renderOnce: () => Promise<void>
@@ -35,28 +35,6 @@ afterEach(async () => {
     currentRenderer.destroy()
   }
 })
-
-// Settle Diff highlighting deterministically. Each iteration:
-// 1. Render twice — the first render may trigger Diff.requestRebuild via microtask
-//    (runs during renderOnce's internal awaits), which calls requestRender while
-//    rendering=true, setting immediateRerenderRequested. The resulting re-render
-//    is scheduled via clock.setTimeout (ManualClock), so needs a second renderOnce.
-// 2. Resolve all pending highlights (proper signal via mock)
-// 3. Await Code.highlightingDone on both sides (proper signal from Code)
-// Loop exits when mock has no more pending requests (state-based).
-async function settleDiffHighlighting(diff: DiffRenderable) {
-  const MAX = 15
-  for (let i = 0; i < MAX; i++) {
-    await renderOnce()
-    await renderOnce()
-    if (!mockClient.isHighlighting()) break
-    mockClient.resolveAllHighlightOnce()
-    const left: CodeRenderable | null = (diff as any).leftCodeRenderable
-    const right: CodeRenderable | null = (diff as any).rightCodeRenderable
-    if (left) await left.highlightingDone
-    if (right) await right.highlightingDone
-  }
-}
 
 // When highlights conceal formatting characters (like **), line lengths change,
 // potentially triggering wrapping changes, height changes, and onResize.
@@ -110,7 +88,7 @@ test("DiffRenderable - no endless loop when concealing markdown formatting", asy
   await renderOnce()
   diffRenderable.wrapMode = "word"
 
-  await settleDiffHighlighting(diffRenderable)
+  await settleDiffHighlighting(diffRenderable, mockClient, renderOnce)
 
   const stats = currentRenderer.getStats()
   expect(stats.frameCount).toBeLessThan(25)
@@ -180,7 +158,7 @@ test("DiffRenderable - line number alignment and gutter heights in split view wi
 
   // First wrapMode toggle: none → word
   diffRenderable.wrapMode = "word"
-  await settleDiffHighlighting(diffRenderable)
+  await settleDiffHighlighting(diffRenderable, mockClient, renderOnce)
   const splitWrapFrame = captureFrame()
 
   const diffChildren = diffRenderable.getChildren()
@@ -220,7 +198,7 @@ test("DiffRenderable - line number alignment and gutter heights in split view wi
   diffRenderable.wrapMode = "none"
   await renderOnce()
   diffRenderable.wrapMode = "word"
-  await settleDiffHighlighting(diffRenderable)
+  await settleDiffHighlighting(diffRenderable, mockClient, renderOnce)
   const splitWrapFrame2 = captureFrame()
   const lines2 = splitWrapFrame2.split("\n")
   let leftLine2Row2 = -1
